[SD][SKV-627] skv2.4适配kerberos认证

+++ run.sh
@@ -1809,7 +1809,10 @@ function run_shell()
     ./pegasus_shell ${CONFIG} $CLUSTER_NAME
     # because pegasus shell will catch 'Ctrl-C' signal, so the following commands will be executed
     # even user inputs 'Ctrl-C', so that the temporary config file will be cleared when exit shell.
-    rm -f ${CONFIG}
+    # however, if it is the specified config file, do not delete it.
+    if [ ${CONFIG_SPECIFIED} -eq 0 ]; then
+        rm -f ${CONFIG}
+    fi
 }
 
 #####################
@@ -1820,6 +1823,7 @@ function usage_migrate_node()
     echo "Options for subcommand 'migrate_node':"
     echo "   -h|--help            print the help info"
     echo "   -c|--cluster <str>   cluster meta lists"
+    echo "   -f|--config <str>    shell config path"
     echo "   -n|--node <str>      the node to migrate primary replicas out, should be ip:port"
     echo "   -a|--app <str>       the app to migrate primary replicas out, if not set, means migrate all apps"
     echo "   -t|--type <str>      type: test or run, default is test"
@@ -1828,6 +1832,7 @@ function usage_migrate_node()
 function run_migrate_node()
 {
     CLUSTER=""
+    CONFIG=""
     NODE=""
     APP="*"
     TYPE="test"
@@ -1840,6 +1845,10 @@ function run_migrate_node()
                 ;;
             -c|--cluster)
                 CLUSTER="$2"
+                shift
+                ;;
+            -f|--config)
+                CONFIG="$2"
                 shift
                 ;;
             -n|--node)
@@ -1864,7 +1873,7 @@ function run_migrate_node()
         shift
     done
 
-    if [ "$CLUSTER" == "" ]; then
+    if [ "$CLUSTER" == "" -a "$CONFIG" == "" ]; then
         echo "ERROR: no cluster specified"
         echo
         usage_migrate_node
@@ -1885,14 +1894,22 @@ function run_migrate_node()
         exit 1
     fi
 
-    echo "CLUSTER=$CLUSTER"
+    if [ "$CLUSTER" != "" ]; then
+        echo "CLUSTER=$CLUSTER"
+    else
+        echo "CONFIG=$CONFIG"
+    fi
     echo "NODE=$NODE"
     echo "APP=$APP"
     echo "TYPE=$TYPE"
     echo
     cd ${ROOT}
     echo "------------------------------"
-    ./scripts/migrate_node.sh $CLUSTER $NODE "$APP" $TYPE
+    if [ "$CLUSTER" != "" ]; then
+        ./scripts/migrate_node.sh $CLUSTER $NODE "$APP" $TYPE
+    else
+        ./scripts/migrate_node.sh $CONFIG $NODE "$APP" $TYPE -f
+    fi
     echo "------------------------------"
     echo
     if [ "$TYPE" == "test" ]; then
@@ -1914,6 +1931,7 @@ function usage_downgrade_node()
     echo "Options for subcommand 'downgrade_node':"
     echo "   -h|--help            print the help info"
     echo "   -c|--cluster <str>   cluster meta lists"
+    echo "   -f|--config <str>    config file path"
     echo "   -n|--node <str>      the node to downgrade replicas, should be ip:port"
     echo "   -a|--app <str>       the app to downgrade replicas, if not set, means downgrade all apps"
     echo "   -t|--type <str>      type: test or run, default is test"
@@ -1922,6 +1940,7 @@ function usage_downgrade_node()
 function run_downgrade_node()
 {
     CLUSTER=""
+    CONFIG=""
     NODE=""
     APP="*"
     TYPE="test"
@@ -1934,6 +1953,10 @@ function run_downgrade_node()
                 ;;
             -c|--cluster)
                 CLUSTER="$2"
+                shift
+                ;;
+            -f|--config)
+                CONFIG="$2"
                 shift
                 ;;
             -n|--node)
@@ -1958,7 +1981,7 @@ function run_downgrade_node()
         shift
     done
 
-    if [ "$CLUSTER" == "" ]; then
+    if [ "$CLUSTER" == "" -a "$CONFIG" == "" ]; then
         echo "ERROR: no cluster specified"
         echo
         usage_downgrade_node
@@ -1979,14 +2002,22 @@ function run_downgrade_node()
         exit 1
     fi
 
-    echo "CLUSTER=$CLUSTER"
+    if [ "$CLUSTER" != "" ]; then
+        echo "CLUSTER=$CLUSTER"
+    else
+        echo "CONFIG=$CONFIG"
+    fi
     echo "NODE=$NODE"
     echo "APP=$APP"
     echo "TYPE=$TYPE"
     echo
     cd ${ROOT}
     echo "------------------------------"
-    ./scripts/downgrade_node.sh $CLUSTER $NODE "$APP" $TYPE
+    if [ "$CLUSTER" != "" ]; then
+        ./scripts/downgrade_node.sh $CLUSTER $NODE "$APP" $TYPE
+    else
+        ./scripts/downgrade_node.sh $CONFIG $NODE "$APP" $TYPE -f
+    fi
     echo "------------------------------"
     echo
     if [ "$TYPE" == "test" ]; then
